from recovery

+++ app/src/main/java/ru/mail/park/diskn/Models/Disk.java
@@ -3,6 +3,8 @@ package ru.mail.park.diskn.Models;
 
 import com.google.gson.annotations.SerializedName;
 
+import java.math.BigDecimal;
+
 
 /**
  *        max_file_size (integer, optional): <Максимальный поддерживаемый размер файла.>,
@@ -11,7 +13,7 @@ import com.google.gson.annotations.SerializedName;
  *  used  trash_size (integer, optional): <Общий размер файлов в Корзине (байт). Входит в used_space.>,
  *        is_paid (boolean, optional): <Признак наличия купленного места.>,
  *  used  used_space (integer, optional): <Используемый объем диска (байт)>,
- *  used  system_folders (SystemFolders, optional): <Адреса системных папок в Диске пользователя.>,
+ *        system_folders (SystemFolders, optional): <Адреса системных папок в Диске пользователя.>,
  *  used  user (User, optional): <Владелец Диска>,
  *        revision (integer, optional): <Текущая ревизия Диска>
  */
@@ -20,20 +22,18 @@ public class Disk {
     @SerializedName("user")
     private final User user;
     @SerializedName("used_space")
-    private final Integer usedSpace;
+    private final BigDecimal usedSpace;
     @SerializedName("total_space")
-    private final Integer totalSpace;
+    private final BigDecimal totalSpace;
     @SerializedName("trash_size")
-    private final Integer trashSize;
-//    @SerializedName("system_folders")
-//    private final SystemFolders systemFolders;
+    private final BigDecimal trashSize;
 
-    public Disk(User user, Integer usedSpace, Integer totalSpace, Integer trashSize) {
+
+    public Disk(User user, BigDecimal usedSpace, BigDecimal totalSpace, BigDecimal trashSize) {
         this.user = user;
         this.usedSpace = usedSpace;
         this.totalSpace = totalSpace;
         this.trashSize = trashSize;
-       // this.systemFolders = systemFolders;
     }
 
 
@@ -41,15 +41,15 @@ public class Disk {
         return user;
     }
 
-    public Integer getUsedSpace() {
+    public BigDecimal getUsedSpace() {
         return usedSpace;
     }
 
-    public Integer getTotalSpace() {
+    public BigDecimal getTotalSpace() {
         return totalSpace;
     }
 
-    public Integer getTrashSize() {
+    public BigDecimal getTrashSize() {
         return trashSize;
     }
 
@@ -63,18 +63,4 @@ public class Disk {
                 '}';
     }
 
-//    public SystemFolders getSystemFolders() {
-//        return systemFolders;
-//    }
-
-//    @Override
-//    public String toString() {
-//        return "Disk{" +
-//                "user=" + user +
-//                ", usedSpace=" + usedSpace +
-//                ", totalSpace=" + totalSpace +
-//                ", trashSize=" + trashSize +
-//                ", systemFolders=" + systemFolders +
-//                '}';
-//    }
 }
